Add GroupMe Bot API
Tying in the GroupMe Bot API call to message.Message._send_response.
This will post a message to the group associated with the Bot ID.
Note that it is required to have a Settings object in the datastore
that has the association between the group ID (given in the initial
callback) and the bot ID (considered secret).