Add ability to customize bot name
If you want to make a bot name different than C-3PO, you should
be able to. This involves customizing the official bot name coming
from the provider (to make sure it doesn't respond to itself) and
the 'mentioned' criteria (what others use to call attention to the
bot).

Fix #69